solc < 0.4.10 trips over long --libraries argument (AN-297)
solc first does a stat on the --libraries argument, to decide whether it
is a file or not. If the string it too long, stat will return ENAMETOOLONG
and solc fails with the error:

terminate called after throwing an instance of 'boost::filesystem::filesystem_error'
  what():  boost::filesystem::status: File name too long: "Library1:0CC835E75E6E3DD60195F2AF69B4AA9BEFFA6D2B  Library2:0CC835E75E6E3DD60195F2AF69B4AA9BEFFA6D2B  Library3:0CC835E75E6E3DD60195F2AF69B4AA9BEFFA6D2B  Library4:0CC835E75E6E3DD60195F2AF69B4AA9BEFFA6D2B  Library5:0CC835E75E6E3DD60195F2AF69B4AA9BEFFA6D2B  Library6:0CC835E75E6E3DD60195F2AF69B4AA9BEFFA6D2B"

This is fixed in solc commit (in v0.4.10 and newer):

https://github.com/ethereum/solidity/commit/14196f2621cbc3a7580ae05b6f7508c92467168a#diff-9f5c1afb3e488e90c0c39502afee1e3c

Work around the problem for earlier versions of solc.